Change the order of registration steps in registrationHandler() in Register.jsx
Now the app first tries registering the user with Firebase authentication service and only after that checks if the username, email address and phone number already exist in database. If they do, the user is deleted from Firebase authentication service and the registration is marked as failed. The order is changed due to necessity to limit read access to users collection to authenticated users only.